requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
gunicorn==21.2.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5 